ModInputSystem At JJP, we had a system that I decided to transfer to Union so that in the future we wouldn't have to create a unique abilities menu for each mod, but could simply use one for all modifications. And also the Input system—it's better to have one library handle events than to have each mod do it separately. In short, if you have created an ability (component) that hangs on a limb, you can easily add it to the abilities menu. To do this, simply implement the IShowInAbilitiesMenu interface, for example MyAbility : MonoBehaviour, IShowInAbilitiesMenu It will add several methods for the menu, such as sprite icons, text display in the menu, and more. A little about ModInputSystem: in addition to OpenAbilitiesMenu, you can also add an event for the PhysicalBehaviour you specified using the method AddPhysInputAction(PhysicalBehaviour physicalBehaviour, ModInputKey key, UnityAction action) And also remove it RemovePhysInputAction(PhysicalBehaviour physicalBehaviour, ModInputKey key, UnityAction action)
ModInputKey is linked to Controls. If the player has changed AltActiveKeyIDENSYS from H to J, for example, you don't need to worry, and in the code it will also be H, because the keycode is linked to the identifier. For convenience in development, we simply remember these keycodes, of which there are only three
AltActiveKeyIDENSYS (from the ENSYS library; if you only use Union, this control will be created by Union itself), default H
union_2, default J
union_3, default K
And also union_open_amenu, which opens the abilities menu, default Y

We also fixed something else, cut out some debug logs that we accidentally left, but that's all, thanks for reading this.